Remove 'generic_ramdisk' and add 'boot' certification
* Remove 'generic_ramdisk' certification since we are no longer
  certifying it in VTS.
* Add 'boot' certification which certifies the entire boot.img.
* 'generic_kernel' certification certifies the kernel.img.
* Remove --boot_signature option.
* Change Android T GKI to have fixed 16K boot_signature size, and must
  be appended at the end of partition.

Bug: 211741246
Test: atest --host mkbootimg_test
Test: atest --host retrofit_gki_test
Test: ./boot_signature_info.sh boot-5.10.img

+++ gki/boot_signature_info.sh
@@ -74,7 +74,7 @@ while IFS= read -r -d '' ARG; do
 done < "${BOOT_IMAGE_ARGS}"
 
 BOOT_IMAGE_VERSION="$(get_arg --header_version "${boot_args[@]}")"
-if [[ "${BOOT_IMAGE_VERSION}" -ge 4 ]]; then
+if [[ "${BOOT_IMAGE_VERSION}" -ge 4 ]] && [[ -f "${BOOT_IMAGE_DIR}/boot_signature" ]]; then
   cp "${BOOT_IMAGE_DIR}/boot_signature" "${VBMETA_IMAGE}"
 else
   tail -c "$(( 16 << 10 ))" "${BOOT_IMAGE}" > "${VBMETA_IMAGE}"
